+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Change `ReplicantSubscription.getId()` to return a primitive integer rather than a boxed `Integer`.
+
 ### [v6.12](https://github.com/realityforge/replicant/tree/v6.12) (2018-09-27)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.11...v6.12)